No more including the terminal
Can you believe this absolute travesty called #include "terminal.h" being removed? I can't believe Landar would ever do something as inherently evil and disgusting as removing the include of the terminal? It may not have been neccessary but when does that matter at the end of the day? Think about the ceremonial monarchy in many countries; what point is there to its existence? None, you may think. But they still exist to this day. Alas, the #include of the terminal header in the chapter 1b

This is a very sad day for GitHub as a whole and I wouldn't be surprised if they terminated Landar's account for this.